Updated validation scripts.
