+++ scripts/ci/validate-module-re-exports.mjs
@@ -9,8 +9,7 @@ import { publicTree } from './exports/policy/public-tree.mjs';
 import { isCommandClass } from './exports/policy/is-command.mjs';
 import { isPluginClass } from './exports/policy/is-plugin.mjs';
 import { isEvent } from './exports/policy/is-event.mjs';
-import { validateCommandExports } from './exports/policy/validate-command-exports.mjs';
-import { validatePluginExports } from './exports/policy/validate-plugin-exports.mjs';
+import { validateClassExports } from './exports/policy/validate-class-exports.mjs';
 import { Export } from './exports/utils/export.mjs';
 import { logData, mapper } from './exports/utils/logger.mjs';
 import { packageNameFromFileName } from './exports/utils/misc.mjs';
@@ -43,14 +42,12 @@ async function main() {
 	const exportsToFix = getExportsToFix( library );
 	const declarationsWithMissingExports = getDeclarationsWithMissingExports( library );
 	const declarationsReferencingInternals = getDeclarationsReferencingInternals( library );
-	const commandExportErrors = validateCommandExports( library );
-	const pluginExportErrors = validatePluginExports( library );
+	const classExportErrors = validateClassExports( library );
 	const dataToLogUnwrapped = [
 		...declarationsWithMissingExports,
 		...declarationsReferencingInternals,
 		...exportsToFix,
-		...commandExportErrors,
-		...pluginExportErrors
+		...classExportErrors
 	];
 
 	// Do not log exceptions that are expected as errors.
